pyproject.toml
[project]
name = "voice-llm"
version = "0.1.0"
description = "Voice AI bot built with Pipecat"
requires-python = ">=3.10"
dependencies = [
    "pipecat-ai[cartesia,daily,google,openai,gladia,local-smart-turn-v3,runner,silero,webrtc]",
    "pipecat-ai-whisker",
    "pipecat-ai-tail",
    "langchain",
    "langchain-community",
    "langchain-openai",
    "langchain-chroma",
    "chromadb",
    "python-docx",
    "pydantic",
    "docx2txt",
    "httpx",
    "chardet",
]

[dependency-groups]
dev = [
    "pyright>=1.1.404,<2",
    "ruff>=0.12.11,<1",
]

[tool.ruff]
line-length = 100
[tool.ruff.lint]
select = ["I"]